requirements.txt
openai>=1.30.0
langgraph>=0.0.20
gradio>=4.16.0
pydantic>=2.5.0
pydantic-core>=2.14.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
